feat: 타입 선언 위치 변경
