--- apps/web/src/entities/auth/model/enrollment-verification/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
