+++ apps/web/src/entities/auth/model/index.ts
@@ -1,4 +1,4 @@
 export * from './form';
 export * from './store';
-export * from './enrollment-verification';
 export * from './types';
+export * from './queries';
